requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
openpyxl>=3.1.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
